# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.PostUrlBody:
    properties:
      interval:
        type: integer
      url:
        type: string
    type: object
  handler.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.UserSignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  repo.CreateURLRow:
    properties:
      id:
        type: string
      intervalSeconds:
        format: int32
        type: integer
      url:
        type: string
    type: object
  repo.CreateUserRow:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  repo.ListURLsByUserRow:
    properties:
      intervalSeconds:
        format: int32
        type: integer
      url:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for monitoring website uptime.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Uptime Monitor API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and set cookie
      parameters:
      - description: Login Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "login successful"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid json format
          schema:
            type: string
        "401":
          description: invalid email or password
          schema:
            type: string
        "500":
          description: error while creating token / json parsing failed
          schema:
            type: string
      summary: User Login
      tags:
      - user
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Sign Up Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserSignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.CreateUserRow'
        "400":
          description: Invalid json request
          schema:
            type: string
        "409":
          description: User already exists
          schema:
            type: string
        "500":
          description: Internal Server error
          schema:
            type: string
      summary: User SignUp
      tags:
      - user
  /urls:
    get:
      description: Get all monitored URLs for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repo.ListURLsByUserRow'
            type: array
        "401":
          description: username not found
          schema:
            type: string
        "404":
          description: url not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Get URLs
      tags:
      - urls
    post:
      consumes:
      - application/json
      description: Add a new URL to monitor for the authenticated user
      parameters:
      - description: Add URL Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.PostUrlBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.CreateURLRow'
        "400":
          description: Invalid json input / interval should be greater than zero.
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Post URL
      tags:
      - urls
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"